requirements.txt
scapy==2.6.1
fastapi==0.115.8
uvicorn[standard]==0.34.0
asyncpg==0.30.0
PyYAML==6.0.2
aiohttp==3.11.13
apscheduler==3.10.4
websockets==14.2
geoip2==4.8.1
python-dotenv==1.0.1
PyJWT==2.10.1
bcrypt>=4.0.0
prometheus-client==0.21.1
ruamel.yaml>=0.18.0
alembic>=1.13.0
psycopg2-binary>=2.9.0
redis[hiredis]>=5.0.0
structlog>=24.0.0
